# api/openapi.yaml
openapi: 3.1.0
info:
  title: Greetd API
  description: A friendly greeting and message management API
  version: 1.0.0
  contact:
    name: Greetd API Support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
paths:
...
  /message:
    get:
      summary: Get the current stored message
      description: Retrieves the currently stored message
      operationId: getMessage
      responses:
        '200':
          description: Current message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: "Hello, World!"

    post:
      summary: Update the stored message
      description: Updates the message that is persisted to disk
      operationId: setMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageRequest'
            example:
              message: "Hello, Universe!"
      responses:
        '200':
          description: Message updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: "Hello, Universe!"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Message cannot be empty"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Failed to save message"
...
components:
  schemas:
...
    MessageRequest:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Message to store
          minLength: 1
          example: "Hello, Universe!"
    MessageResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Stored message
          example: "Hello, World!"
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: "Invalid input"